standardized header into a separate component with its own properties
