+++ src/pages/our-team.js
@@ -1,14 +1,17 @@
 import React from "react"
 import { Link } from "gatsby"
-
 import Layout from "../components/layout/layout"
 import SEO from "../components/seo"
-import NavBar from "../components/nav-bar/nav-bar"
+import Header from "../components/header/header"
 
 const TeamPage = () => (
-  <Layout>
-    <NavBar></NavBar>
+  <Layout navbarStyle="gradient">
     <SEO title="Team Page" />
+    <Header
+      title="Meet the team"
+      subheading="Learn a little about ViTAL key members"
+      imageBool={false}
+    />
     <h1>Our Team</h1>
     <p>Welcome to page 2</p>
     <Link to="/">Go back to the homepage</Link>
